Extract License Requests & Submissions views tests
